requirements.txt
fastapi
uvicorn
pydantic
python-docx
pypandoc
pillow
sqlalchemy
pytest
aiofiles
python-multipart